Refactor exposed API as funcs
Making use of init() was never a good idea in the first place. It's
better to let the user decide what regex to compile and when.

Most importantly, we will start making the regexes configurable in
certain ways, so we need funcs.

Also remove the benchmarks, since the configurability will make them
obsolete.