don't allow | as part of a URL
This was an unintended bug added ages ago. In c64477d7d2, we modified
midChar to add \p{Po} to the character set. Unfortunately, we added |,
essentially ending up with [somechars|\p{Po}]. In other words, | was now
part of the set, while we only meant it as a regex "or" operator.

Fix this, and add a test.

Fixes #26.