don't allow underscores and asterisks as end chars
This way, we match URLs within markdown emphasis like *this* or _this_
without the trailing character.

It's highly unlikely that URLs would genuinely want to end in either of
those characters. Asterisks are reserved, so they shouldn't be in the
path to begin with. And underscores are most common in the middle of the
path, and are already common to style plaintext at the start or end of
URLs.

Still allow them in the middle of URLs, because there's no reason not
to.

Fixes #29.