stop using \b to catch the end of TLDs in Relaxed
This only worked for ASCII TLDs, since \b is a zero-length separator
between \w and \W (word and non-word), and \w only accounts for ASCII
letters.

We could use \b exclusively for ASCII TLDs, but that would complicate
and slow down the regex, and result in an unfair advantage and
inconsistency versus non-ASCII TLDs.

We don't have any alternative for \b either, such as a word-separator
escape code that's Unicode-aware.

Because of the reasons above, just drop this feature entirely. If any
user notices and complains about it, we can reconsider our options.

Fixes #32.