generate: make tldsgen non-racy
The waitgroup was almost right, but we only waited for the http GETs to
finish, not for the map writes to finish.

Remove the extra goroutine and write straight to the map to fix the
issue. And error handling via panics is fine for a small code generator
program like this one.

Fixes #44.