add a list of unofficial schemes
This is mainly to match URLs for some pieces of software, like
PostgreSQL and Slack.

There is no list we can automatically pull from, because these are all
unofficial and not registered.

It seems like most software is gradually switching to shared prefixes
like jdbc or just https urls with host matching, but some are still
relatively widespread.

Fixes #42.