improve benchmarks to be more realistic
First, include benchmarks for StrictMatchingScheme.
Second, make the benchmarks call the API again and again,
which is the worst-case scenario we use sync.Once for.
Third, report allocs and speed.

Finally, tweak the inputs to be a bit longer.
We also remove the "single match" input,
as it doesn't really contribute anything on its own.

Note that the StrictMatchingScheme benchmarks don't call the API repeatedly.
This is on purpose, as otherwise we'd just benchmark regexp.Compile.
In practice, each use of StrictMatchingScheme will be different,
and so we can't do global caching like with the others.
We expect the user to be careful to reuse the result.

	name                          time/op
	Strict_none-16                  23.0µs ± 1%
	Strict_many-16                  19.8µs ± 2%
	Relaxed_none-16                 31.4µs ± 1%
	Relaxed_many-16                 28.0µs ± 1%
	StrictMatchingScheme_none-16     514ns ± 2%
	StrictMatchingScheme_many-16    1.01µs ± 0%

	name                          speed
	Strict_none-16                3.39MB/s ± 1%
	Strict_many-16                4.89MB/s ± 2%
	Relaxed_none-16               2.48MB/s ± 1%
	Relaxed_many-16               3.47MB/s ± 2%
	StrictMatchingScheme_none-16   152MB/s ± 2%
	StrictMatchingScheme_many-16  96.1MB/s ± 0%

	name                          alloc/op
	Strict_none-16                    164B ± 5%
	Strict_many-16                    398B ± 2%
	Relaxed_none-16                   160B ± 0%
	Relaxed_many-16                   730B ± 1%
	StrictMatchingScheme_none-16     0.00B
	StrictMatchingScheme_many-16      228B ± 0%

	name                          allocs/op
	Strict_none-16                    1.00 ± 0%
	Strict_many-16                    6.00 ± 0%
	Relaxed_none-16                   1.00 ± 0%
	Relaxed_many-16                   7.00 ± 0%
	StrictMatchingScheme_none-16      0.00
	StrictMatchingScheme_many-16      3.00 ± 0%